refactor: ddd interface
